refactor: move createpayableFromTransaction use case to its own folder
