+++ src/modules/payables/payables.module.ts
@@ -3,7 +3,7 @@ import { PayablesController } from './payables.controller';
 import { MerchantPayable } from 'src/shared/domain/models';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { MerchantPayableRepository } from './infra/repositories/merchant-payable.repository';
-import { CreatePayableFromTransactionUseCase } from './application/usecases/create-payable-from-transaction.usecase';
+import { CreatePayableFromTransactionUseCase } from './application/usecases/CreatePayableFromTransaction/create-payable-from-transaction.usecase';
 import { TransactionCreatedListener } from './application/listeners/transaction-created.listener';
 
 @Module({
